Create new test_ephemeral and refactor with test_normalization
